pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pathlib>=1.0.1",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.46.0",
]
